+++ hardpy/hardpy_panel/api.py
@@ -2,15 +2,32 @@
 # GNU General Public License v3.0 (see LICENSE or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 import os
+import re
+from enum import Enum
+from urllib.parse import unquote
 
 from fastapi import FastAPI
 from fastapi.staticfiles import StaticFiles
 
-from hardpy.pytest_hardpy.utils import ConfigData, RunStatus as Status
+from hardpy.common.config import ConfigManager
 from hardpy.pytest_hardpy.pytest_wrapper import PyTestWrapper
 
 app = FastAPI()
 app.state.pytest_wrp = PyTestWrapper()
+
+
+class Status(str, Enum):  # noqa: WPS600
+    """Pytest run status.
+
+    Statuses, that can be returned by HardPy to frontend.
+    """
+
+    STOPPED = "stopped"
+    STARTED = "started"
+    COLLECTED = "collected"
+    BUSY = "busy"
+    READY = "ready"
+    ERROR = "error"
 
 
 @app.get("/api/start")
@@ -55,12 +72,12 @@ def couch_connection():
     """Get couchdb connection string.
 
     Returns:
-        dict[str, str, str]: couchdb connection string
+        dict[str, str]: couchdb connection string
     """
-    config_data = ConfigData()
+    connection_url = ConfigManager().get_config().database.connection_url()
 
     return {
-        "connection_str": config_data.connection_string,
+        "connection_str": connection_url,
     }
 
 
@@ -74,13 +91,39 @@ def confirm_dialog_box(dialog_box_output: str):
     Returns:
         dict[str, RunStatus]: run status
     """
-    if app.state.pytest_wrp.confirm_dialog_box(dialog_box_output):
+    hex_base = 16
+    unquoted_string = unquote(dialog_box_output)
+    decoded_string = re.sub(
+        "%([0-9a-fA-F]{2})",
+        lambda match: chr(int(match.group(1), hex_base)),
+        unquoted_string,
+    )
+
+    if app.state.pytest_wrp.send_data(str(decoded_string)):
+        return {"status": Status.BUSY}
+    return {"status": Status.ERROR}
+
+
+@app.post("/api/confirm_operator_msg/{is_msg_visible}")
+def confirm_operator_msg(is_msg_visible: bool):
+    """Confirm operator msg.
+
+    Args:
+        is_msg_visible (bool): is operator message is visible
+
+    Returns:
+        dict[str, RunStatus]: run status
+    """
+    if app.state.pytest_wrp.send_data(str(is_msg_visible)):
         return {"status": Status.BUSY}
     return {"status": Status.ERROR}
 
 
 app.mount(
     "/",
-    StaticFiles(directory=(os.path.dirname(__file__)) + "/frontend/dist", html=True),
+    StaticFiles(
+        directory=(os.path.dirname(__file__)) + "/frontend/dist",  # noqa: WPS336
+        html=True,
+    ),
     name="static",
 )