+++ hardpy/pytest_hardpy/db/schema.py
@@ -1,8 +1,6 @@
 # Copyright (c) 2024 Everypin
 # GNU General Public License v3.0 (see LICENSE or https://www.gnu.org/licenses/gpl-3.0.txt)
 
-from typing import Optional
-
 from pydantic import BaseModel, Field, ConfigDict
 
 from hardpy.pytest_hardpy.utils import TestStatus as Status
@@ -30,6 +28,7 @@ class CaseStateStore(IBaseResult):
         "stop_time": 1695817189,
         "assertion_msg": null,
         "msg": null,
+        "attempt": 1,
         "dialog_box": {
           "title_bar": "Example of text input",
           "dialog_text": "Type some text and press the Confirm button",
@@ -46,6 +45,7 @@ class CaseStateStore(IBaseResult):
     assertion_msg: str | None = None
     msg: dict | None = None
     dialog_box: dict = {}
+    attempt: int = 0
 
 
 class CaseRunStore(IBaseResult):
@@ -59,13 +59,15 @@ class CaseRunStore(IBaseResult):
         "stop_time": 1695817189,
         "assertion_msg": null,
         "msg": null,
+        "attempt": 1,
         "artifact": {}
     }
     """
 
     assertion_msg: str | None = None
     msg: dict | None = None
-    artifact: Optional[dict] = {}
+    artifact: dict = {}
+    attempt: int = 0
 
 
 class ModuleStateStore(IBaseResult):
@@ -118,7 +120,7 @@ class ModuleRunStore(IBaseResult):
     """
 
     cases: dict[str, CaseRunStore] = {}
-    artifact: Optional[dict] = {}
+    artifact: dict = {}
 
 
 class Dut(BaseModel):
@@ -127,9 +129,10 @@ class Dut(BaseModel):
     Example:
     "dut": {
         "serial_number": "a9ad8dca-2c64-4df8-a358-c21e832a32e4",
+        "part_number": "part_number_1",
         "info": {
-        "batch": "test_batch",
-        "board_rev": "rev_1"
+          "batch": "test_batch",
+          "board_rev": "rev_1"
         }
     },
     """
@@ -137,6 +140,25 @@ class Dut(BaseModel):
     model_config = ConfigDict(extra="forbid")
 
     serial_number: str | None
+    part_number: str | None
+    info: dict = {}
+
+
+class TestStand(BaseModel):
+    """Test stand description.
+
+    Example:
+    "test_stand": {
+        "name": "test_stand_1",
+        "info": {
+          "geo": "Belgrade",
+        }
+    },
+    """
+
+    model_config = ConfigDict(extra="forbid")
+
+    name: str | None
     info: dict = {}
 
 
@@ -158,6 +180,7 @@ class ResultStateStore(IBaseResult):
       "name": "hardpy-stand",
       "dut": {
         "serial_number": "92c5a4bb-ecb0-42c5-89ac-e0caca0919fd",
+        "part_number": "part_1",
         "info": {
           "batch": "test_batch",
           "board_rev": "rev_1"
@@ -165,6 +188,7 @@ class ResultStateStore(IBaseResult):
       },
       "test_stand": {
         "name": "Test stand 1"
+        "info": {}
       },
       "drivers": {
         "driver_1": "driver info",
@@ -172,6 +196,11 @@ class ResultStateStore(IBaseResult):
           "state": "active",
           "port": 8000
         }
+      },
+      "operator_msg": {
+        "msg": "Operator message",
+        "title": "Message",
+        "visible": "True"
       },
       "modules": {
         "test_1_a": {
@@ -187,6 +216,7 @@ class ResultStateStore(IBaseResult):
               "stop_time": 1695817264,
               "assertion_msg": null,
               "msg": null,
+              "attempt": 1,
               "dialog_box": {
                 "title_bar": "Example of text input",
                 "dialog_text": "Type some text and press the Confirm button",
@@ -204,6 +234,7 @@ class ResultStateStore(IBaseResult):
               "start_time": 1695817264,
               "stop_time": 1695817264,
               "assertion_msg": "The test failed because minute 21 is odd! Try again!",
+              "attempt": 1,
               "msg": [
                 "Current minute 21"
               ]
@@ -220,10 +251,11 @@ class ResultStateStore(IBaseResult):
     id: str = Field(..., alias="_id")
     progress: int
     timezone: tuple[str, str] | None = None
-    test_stand: dict = {}
+    test_stand: TestStand
     dut: Dut
     modules: dict[str, ModuleStateStore] = {}
-    drivers: Optional[dict] = {}
+    drivers: dict = {}
+    operator_msg: dict = {}
 
 
 class ResultRunStore(IBaseResult):
@@ -244,6 +276,7 @@ class ResultRunStore(IBaseResult):
       "name": "hardpy-stand",
       "dut": {
         "serial_number": "92c5a4bb-ecb0-42c5-89ac-e0caca0919fd",
+        "part_number": "part_1",
         "info": {
           "batch": "test_batch",
           "board_rev": "rev_1"
@@ -251,6 +284,7 @@ class ResultRunStore(IBaseResult):
       },
       "test_stand": {
         "name": "Test stand 1"
+        "info": {}
       },
       "drivers": {
         "driver_1": "driver info",
@@ -275,6 +309,7 @@ class ResultRunStore(IBaseResult):
               "stop_time": 1695817264,
               "assertion_msg": null,
               "msg": null,
+              "attempt": 1,
               "artifact": {}
             },
             "test_minute_parity": {
@@ -286,6 +321,7 @@ class ResultRunStore(IBaseResult):
               "msg": [
                 "Current minute 21"
               ],
+              "attempt": 1,
               "artifact": {
                 "data_str": "123DATA",
                 "data_int": 12345,
@@ -306,8 +342,8 @@ class ResultRunStore(IBaseResult):
     id: str = Field(..., alias="_id")
     progress: int
     timezone: tuple[str, str] | None = None
-    test_stand: dict = {}
+    test_stand: TestStand
     dut: Dut
     modules: dict[str, ModuleRunStore] = {}
-    drivers: Optional[dict] = {}
-    artifact: Optional[dict] = {}
+    drivers: dict = {}
+    artifact: dict = {}