+++ hardpy/pytest_hardpy/utils/__init__.py
@@ -3,11 +3,13 @@
 
 from hardpy.pytest_hardpy.utils.node_info import NodeInfo
 from hardpy.pytest_hardpy.utils.progress_calculator import ProgressCalculator
-from hardpy.pytest_hardpy.utils.const import TestStatus, RunStatus
+from hardpy.pytest_hardpy.utils.const import TestStatus
 from hardpy.pytest_hardpy.utils.singleton import Singleton
-from hardpy.pytest_hardpy.utils.config_data import ConfigData
+from hardpy.pytest_hardpy.utils.connection_data import ConnectionData
 from hardpy.pytest_hardpy.utils.exception import (
     DuplicateSerialNumberError,
+    DuplicatePartNumberError,
+    DuplicateTestStandNameError,
     DuplicateDialogBoxError,
     WidgetInfoError,
 )
@@ -27,10 +29,11 @@ __all__ = [
     "NodeInfo",
     "ProgressCalculator",
     "TestStatus",
-    "RunStatus",
     "Singleton",
-    "ConfigData",
+    "ConnectionData",
     "DuplicateSerialNumberError",
+    "DuplicatePartNumberError",
+    "DuplicateTestStandNameError",
     "DuplicateDialogBoxError",
     "WidgetInfoError",
     "DialogBox",