+++ hardpy/pytest_hardpy/utils/singleton.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2024 Everypin
 # GNU General Public License v3.0 (see LICENSE or https://www.gnu.org/licenses/gpl-3.0.txt)
 
-class Singleton(object):
+class Singleton:
     """Singleton class.
 
     In the child class must be used constructor of type: