+++ docs/changelog.md
@@ -2,6 +2,29 @@
 
 Versions follow [Semantic Versioning](https://semver.org/): `<major>.<minor>.<patch>`.
 
+## HardPy 0.6.0
+
+In HardPy, compared to version 0.5.0 and less, the startup principle has changed.
+The `hardpy-panel` command is no longer available.
+
+The HardPy project from version 0.6.0 onwards must contain the **hardpy.toml** file.
+
+* Add clearing the **statestore** database before running hardpy using
+  the `--hardpy-clear-database` option of the **pytest-hardpy** plugin.
+* Add the **name** and **info** fields to **test_stand** in the database schema.
+* Add the **part_number** field to **dut** in the database schema.
+* Add the **attempt** field to test case in the database schema.
+* Add `set_stand_name` and `set_dut_part_number` functions.
+* Add a hardpy template project using the `hardpy init` command.
+* Add hardpy config .toml file - **hardpy.toml**.
+* Refactor **pytest-hardpy** plugin options.
+* Add CLI to hardpy as an entry point. The `hardpy-panel` command is now unavailable.
+* Fix using special characters in dialog boxes. ASCII symbols are passed from frontend to backend.
+* Add unit tests for dialog widgets.
+* Fix statuses of stopped tests.
+* Fix progress bar with skipping tests. Progress bar fills to the end when tests are skipped.
+* Add report name generation when serial number is missing.
+
 ## HardPy 0.5.1
 
 * Add the ability to work with cloud couchdb via couchdb config.